+++ src/ordhook/server.ts
@@ -14,11 +14,11 @@ export const ORDHOOK_BASE_PATH = `http://${ENV.CHAINHOOK_NODE_RPC_HOST}:${ENV.CH
 export const PREDICATE_UUID = randomUUID();
 
 /**
- * Starts the chainhooks event server.
+ * Starts the Ordhook event observer.
  * @param args - DB
  * @returns ChainhookEventObserver instance
  */
-export async function startChainhookServer(args: { db: PgStore }): Promise<ChainhookEventObserver> {
+export async function startOrdhookServer(args: { db: PgStore }): Promise<ChainhookEventObserver> {
   const predicates: ServerPredicate[] = [];
   if (ENV.CHAINHOOK_AUTO_PREDICATE_REGISTRATION) {
     const blockHeight = await args.db.getChainTipBlockHeight();
@@ -54,7 +54,8 @@ export async function startChainhookServer(args: { db: PgStore }): Promise<Chain
     base_url: ORDHOOK_BASE_PATH,
   };
   const server = new ChainhookEventObserver(serverOpts, ordhookOpts);
-  await server.start(predicates, async (_uuid: string, payload: Payload) => {
+  await server.start(predicates, async (uuid: string, payload: Payload) => {
+    logger.info(`OrdhookServer received payload from predicate ${uuid}`);
     await args.db.updateInscriptions(payload);
   });
   return server;
